+++ v3/crates/compatibility/src/compatibility_date.rs
@@ -155,7 +155,8 @@ pub fn get_compatibility_date_for_flag(flag: Flag) -> Option<CompatibilityDate> 
         Flag::DisallowDuplicateAggregateFunctionDefinitionsForScalarType => {
             Some(new_compatibility_date(2025, 1, 25))
         }
-        Flag::DisallowDataConnectorScalarTypesMismatch
+        Flag::TypecheckObjectTypeValuesInPresets
+        | Flag::DisallowDataConnectorScalarTypesMismatch
         | Flag::CheckObjectTypeFieldsExist
         | Flag::DisallowOrderByFieldsWithFieldArguments => Some(new_compatibility_date(2025, 2, 4)),
     }