+++ v3/crates/open-dds/src/flags.rs
@@ -37,6 +37,7 @@ pub enum Flag {
     DisallowMultidimensionalArraysInBooleanExpressions,
     DisallowDuplicateNamesAcrossTypesAndExpressions,
     DisallowDuplicateAggregateFunctionDefinitionsForScalarType,
+    TypecheckObjectTypeValuesInPresets,
     DisallowDataConnectorScalarTypesMismatch,
     CheckObjectTypeFieldsExist,
     DisallowOrderByFieldsWithFieldArguments,
@@ -100,6 +101,7 @@ impl Flag {
             Flag::DisallowDuplicateAggregateFunctionDefinitionsForScalarType => {
                 "disallow_duplicate_aggregate_function_definitions_for_scalar_type"
             }
+            Flag::TypecheckObjectTypeValuesInPresets => "typecheck_object_type_values_in_presets",
             Flag::DisallowDataConnectorScalarTypesMismatch => {
                 "disallow_data_connector_scalar_types_mismatch"
             }